Damian

+++ internal/model/comment.go
@@ -3,8 +3,8 @@ package model
 import "database/sql"
 
 type Comment struct {
-	ID         int            `json:"id"`
-	Task       int            `json:"id_task"`
+	ID         int64          `json:"id"`
+	Task       int64          `json:"id_task"`
 	Text       string         `json:"text"`
 	UserName   string         `json:"userName"`
 	Time       string         `json:"time"`
